Applied nested routing to all needed components, deleted modal-based auth components
