+++ catlas/src/pages/index.js
@@ -1,11 +1,6 @@
 import Main from './Main';
 import Clubs from './Clubs';
 import About from './About';
-import Talks from './Talks';
 import Links from './Links';
-import Scenes from './Scenes';
-import Login from './Login';
-import Register from './Register';
-import Reset from './Reset';
 
-export { Main, Clubs, About, Talks, Links, Scenes, Login, Register, Reset };
+export { Main, Clubs, About, Links };
